GCW-3673 Move markAsRead on mount to useMessages
